pyproject.toml
[project]
name = "claude-wake-up"
version = "0.1.0"
description = "Telegram bot to remotely launch Claude Code sessions in Zellij"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []
